// integrations/netsuite-tba/schema.ts
export const netsuiteCustomerCreateInputSchema = z.object({
    externalId: z.string().min(1),
    name: z.string().min(1),
    email: z.string().email(),
    taxNumber: z.string().optional(),
    phone: z.string().optional(),
    addressLine1: z.string().optional(),
    addressLine2: z.string().optional(),
    city: z.string().optional(),
    state: z.string().optional(),
    zip: z.string().optional(),
    country: z.string().optional()
});